requirements.txt
Django==4.2.3
djangorestframework==3.14.0
markdown==3.4.4
django-filter
Pillow==10.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
djangorestframework-simplejwt==5.2.2
django-phonenumber-field[phonenumberslite]
django-extensions
faker
pytest
pytest-django
drf-spectacular
django-cors-headers
llama_index
nltk
llama-cpp-python
django-modeltranslation

black
isort